writer/requirements.txt
kafka-python==2.0.2
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
